Rename: Pagination 컴포넌트 공통 컴포넌트로 구분
